chore: renamed common error to known error
