+++ frontend/pages/404.tsx
@@ -1,12 +1,12 @@
 import router from "next/router";
 
-import { PlainErrorView } from "~frontend/views/ErrorView";
+import { ErrorView } from "~frontend/views/ErrorView";
 import { routes } from "~shared/routes";
 import { Button } from "~ui/buttons/Button";
 
 export default function ErrorPage() {
   return (
-    <PlainErrorView
+    <ErrorView
       title="Page not found"
       description={
         <>
@@ -22,6 +22,6 @@ export default function ErrorPage() {
       >
         Go back to homepage
       </Button>
-    </PlainErrorView>
+    </ErrorView>
   );
 }
