+++ frontend/src/errors/sentryFallbackErrorRenderer.tsx
@@ -1,10 +1,10 @@
 // This type is not nicely exported. As we use import type - it has no runtime artifacts so seems to be safe
 import type { FallbackRender } from "@sentry/react/dist/errorboundary";
 
-import { ErrorView } from "./ErrorView";
+import { SmartErrorView } from "./SmartErrorView";
 
 export const sentryFallbackErrorRenderer: FallbackRender = (errorData) => {
   const { error, resetError } = errorData;
 
-  return <ErrorView error={error} resetError={resetError} />;
+  return <SmartErrorView error={error} resetError={resetError} />;
 };
